requirements.txt
# 核心深度学习框架
torch>=1.13.0
torchvision>=0.14.0

# YOLO相关
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0

# QR码生成
qrcode>=7.0
segno>=1.5.0

# 数据增强
albumentations>=1.3.0

# Web服务
fastapi>=0.95.0
uvicorn>=0.21.0

# ONNX支持
onnx>=1.14.0
onnxruntime>=1.15.0

# Core ML支持 (Mac)
coremltools>=6.0

# 其他工具
matplotlib>=3.5.0
tqdm>=4.64.0
pyyaml>=6.0